type change

+++ src/main/java/com/gomugomu/ma_java_avancee_projet_backend/student/StudentExam.java
@@ -4,12 +4,13 @@ import java.util.UUID;
 
 import com.gomugomu.ma_java_avancee_projet_backend.common.Base;
 
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
 import lombok.Data;
+import jakarta.persistence.Id;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.IdClass;
@@ -32,5 +33,5 @@ public class StudentExam extends Base {
   private UUID studentId;
 
   @Column(nullable = true)
-  private short score;
+  private Float score;
 }
